+++ BeemEnergyAddon/config.yaml
@@ -3,7 +3,7 @@ url: "https://github.com/funkolab/BeemEnergy"
 description: "Connect your Beem Energy Solar Panel to Home Assistant ⚡️"
 services:
   - "mqtt:want"
-version: "0.1.0"
+version: "0.2.0"
 homeassistant_api: true
 image: "ghcr.io/funkolab/beemenergy"
 slug: "beem_energy"
